backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
bcrypt==4.1.1
